Added timeDeriv() function to all fields
timeDeriv() returns a pointer to a variable containing the
time-derivative (if not allocated already then allocates). A macro
ddt is defined in globals.h which allows concise access to this
giving a more readable form:

ddt(var) = ...

The bout_solve functions are now simplified to take just a single field
argument.