Fixed bug in invert_laplace:
The bug is in line 141 of invert_laplace.cpp. In case we don't have the
same line number, it is the line that looks like

coef4 += g11[jx][jy] * 0.25 * ((*ccoef)[jx+1][jy] - (*ccoef)[jx-1][jy])
/ (SQ(dx[jx][jy])*((*ccoef)[jx][jy]));

This is correct now. Previously there was an error there. The error was
that the parentheses in the denominator were wrong so that *ccoef was
being squared, but it should not be, at least according to BOUT06 and to
Maxim's notes.