Added Git revision to log file
When compiling, puts the Git revision number into the
BOUT++ library. When run, this then gets printed out along with
the MD5 checksum as before.

This makes it easier to go back to an old run and get the
version used using "git co <revision>", find differences
between versions etc.