New options handling code. Breaks backwards compatibility
Options is now a class which represents a tree of
options. key-value pairs can be accessed using get
and set methods.

A singleton Options::getRoot() gives access to the
options tree.

To read in options, there is a singleton
OptionsReader::getInstance() which uses objects implementing
the OptionsParser interface.

Implementations of OptionsParser are in src/sys/options/

Currently only got a parser for the current INI file type in
options_ini.cpp

The command line reading is done in OptionsReader::parseCommandLine()
which currently will only set top-level options.