Hypnotoad: Added a manual switch to simplify the boundary
A checkbox on the left of the window "Simple boundary" turns the boundary
into a square box. This is sometimes done by the grid generator automatically
if needed, but this maual setting is useful for generating ITER grids where the upper x-point is very close to the target plates. Using this option then gives more space for the legs.