Changed order of initialisation
Need to be able to put useful error messages in the logs
when reading the options file fails. To open the log files
we need to know the MPI processor number. Therefore MPI
is now initialised at the start, and then PetscInitialise
may be called after the options file has been read in.