Hypnotoad: signs of Bp
Changed sign of Br and Bz (again). Now set so that psi
increasing outwards from core to edge results in Bp
clockwise in the poloidal plane i.e. in the positive
Grad Theta direction:

  Brxy = (dpsi/dz) / Rxy
  Bzxy = -(dpsi/dR) / Rxy

This seems to be the convention for psi in G-EQDSK files.

Added sign of Bp terms to curvature and Jpar expressions,
but not tested yet (no IDL here...)