Hypnotoad: Added option to smooth hthe
Large jumps in hthe close to the x-point can cause
numerical problems. Solution should be to generate a mesh which is
better behaved, but for now hthe can be smoothed in these regions.

To avoid changing hthe in most of the mesh, a nonlinear smoothing
method is used so that smoothing is applied most where the
second derivative of hthe is highest. The code for this is in
smooth_nl.pro