Gridgen: Modified leg_separatrix so works for NSTX case
Code needs to decide if a separatrix is a leg or part of the core. Does this
by taking dot product of line with the line from the x-point to the
primary O-point. In the case of NSTX the x-point is shifted inwards
relative to the o-point. This gave the wrong result when the R-Z mesh
distortion wasn't taken into account.