Added option "openclose" to DataFile
On some machines, occasionally keeping the output file open
all the time results in lost data. By default now the output file
will be opened and closed each time data is written.

This is less than ideal for efficiency, but it's better to
use the safer option as the default. This behavior can be disabled
by setting "openclose = false" in the [io] section.