New physics model API
Rather than implementing C-style functions, physics models
can be defined by defining a class which inherits from
PhysicsModel. This is done by including

<bout/physicsmodel.hxx>

for new-style interface, and

<boutmain.hxx>

for the old style. For now both ways are supported.

An example is in

examples/advect1d-newapi

which is the same as the advect1d example, but using the new API.

Main motivations are:

o To make it easier for large models to be split into multiple files
  without resorting to the preprocessor.

o To enable multiple physics models to coexist. This should make it
  neater to couple two existing models together. Using the old method
  this may have been possible with namespacing, but would be messy.