Small NetCDF fix
NetCDF crashes in strange ways if some characters are used
in variable names. Asterisk (*) seems to do this, so using something like

SAVE_ONCE(*f);

will produce a variable name '*f' and crash NetCDF. To catch this
with a useful error message, the NetCDF wrapper now checks for '*'
in the name, throwing an exception if found.