Adding a few improvements and example routines.
Adding a custom monitor. This replaces the need for "-eps_monitor" and
reports BOUT++ rather than SLEPc eigenvalues.

Adding an example eigenvalue comparison routine to select which of two
eigenvalues we prefer.

Adding the ability to specify real and imaginary targets (in BOUT++
space) which can then be used to set the SLEPc target magnitude
automatically. This is much easier to use than setting target directly
as it doesn't require any changes if we change the timestep etc.