CVODE solver options
Added options to the cvode solver to set the maximum order of the
method (CVodeSetMaxOrd()) and to switch on stability limit
detection in the BDF method (CVodeSetStabLimDet()).

These may be helpful if the solver is trying to take high-order
timesteps and is limited by weakly-damped oscillatory modes: it is
possibly that turning on stability detection or reducing the
maximum order may help (orders 1 and 2 are A-stable and the size of
the instability region increases with order greater than 2).