Call monitors on restart option
Added option (called 'dump_on_restart') which, if true, calls the
monitors when restart=true. This is useful if initial conditions
are set by editing restart files outside of BOUT++. It also means
that the wall_limit is closer to being accurate (otherwise it would
miss counting the first (output) timestep).