Added outine of ParallelTransform classes
Might provide a way to isolate code which is specific to
the Shifted Metric method (zShift etc.), and the FCI scheme.

The idea is that this class takes care of calculating values
along field-lines (the y up down pointers in Field).

Mesh would contain a private pointer to a ParallelTransform class,
passing requests for transforms through from an external interface
(Strategy pattern).