Added ability to read 3D fields in real space
If "nz" is defined in the input grid file, then this fixes
the size of the mesh in Z, and the "MZ" option is ignored.

In this case, 3D variables in the input file are assumed to
be in real space, rather than Fourier transformed. In this
case they must be of the correct size, with nz points in Z.

This closes #22