Bracket handling of staggered fields:
Previously the bracket operator always returned a CELL_CENTRE field.  This would cause problems on when using using the bracket on staggered fields.  Now, bracket will a return a field whose position is the same of the two input fields.  An error will be produced if the two input fields are not in the same location.