Adding FFT wrappers to allow 2d (i)rfft calls.
Note: This is 2d in the sense of doing many 1d FFTs, not a 2d FFT.
Currently only implemented for serial (non-openmp).
May be useful in optimising shiftZ.
Should be relatively easily extensible to higher dimensions.
This has been tested for correctness but not speed. It may be better to
transpose operations (currently the dimension to transform is the last
one, meaning the the related data is strided).