Optimising shiftZ2D slightly.
It seems that the calculation and multiplication of the phase was
dominating the shiftZ2D routine. This commit uses a static cmatrix to
store the precalculated phase factors. This now suggests that shiftZ2D
is about 1/3 of the cost of shiftZ for the test problem, though at the
expense of greater memory consumption (~equivalent to a Field3D).