Adding shiftZ3D approach.
Uses a single call to rfft/irfft per call to shiftZ3D. Timings suggest
this costs exactly the same as shiftZ2D.

Adding cached phase version of shiftZ.

This assumes the phase is always set by +/- ShiftAngle. This is unlikely
to have much impact as it is currently only used for the boundary/guard
regions within boutmesh. This routine could however be adopted by the
other shiftZ wrappers which return a Field3D object.

Adding cached phase for shiftZ to Field3D

This means there should only be one copy of the phase array to be shared
between different shifting routines. Unfortunately we need to retain a
separate phase for the shiftZ (1d) approach as this uses a different
phase shift (used for matching across y-boundaries). I guess we could
have some sort of transform object which contains details like this, so
we call something like Field3D a ; a.transform(shiftedX) ; ...