Python routing for slicing grid files
Select a region of a grid file, and output as a separate file.
Can be used to turn an output from Hypnotoad into a simulation of a
divertor leg (for example).

Added some routines to DataFile to make reading and writing more
convenient and pythonic:

o  keys()   -- same as list()

o __getitem__ and __setitem__ to allow [key] indexing