Adding some alternative ZFFT routines
Initial attempts to optimise the use of ZFFT routines in invert.
Here we precalculate and cache the phase to use. This assumes that
phases is always +/- mesh->zShift, which is currently always true.
As this routine requires different information it is necessary to
replace calls to ZFFT whenever we want to use these routines. Currently
this has only been done for the SPT invert method. It would be possible
to use this for all invert methods, but as there may be further
improvements we'll start by just focusing on the SPT case.