Use datablock access rather than indexing.
This is a small tweak to the Arakawa_Opt bracket method which avoids
using [] operators to access Field data but instead gets a pointer to
the underlying data and then accesses this directly. Essentially
bypasses some assertions -- could be dangerous but as this is an
optimised method we assume this is acknowledged by the user.