Adding cached phase for shiftZ to Field3D
This means there should only be one copy of the phase array to be shared
between different shifting routines. Unfortunately we need to retain a
separate phase for the shiftZ (1d) approach as this uses a different
phase shift (used for matching across y-boundaries). I guess we could
have some sort of transform object which contains details like this, so
we call something like Field3D a ; a.transform(shiftedX) ; ...