Adding initial outline of a routine to shiftz
This updates #46.

It has been tested for a simple case with shiftXDerivs=True,
twistShift=False.

It can be noted that twistShift=True applies shifting to the yGuard
regions only (at least as far as I've found so far). As such it is not
possible to use this shiftZ routine to, e.g. restore the requested
initial conditions from the output field. It's likely I'm missing
something here.