Enable C1 and C2 coefficients in Laplacian solvers
Some Laplacian solvers (in particular PETSc) have setCoefC1 and
setCoefC2 functions to enable a slightly more general equation to
be solved. These are now declared in invert_laplace.hxx so that they
can actually be called from a Laplacian* object and have default
implementations that throw exceptions if they are not overridden by
an implementation.