Adding simple rk4 scheme.
This is a naive implementation of the "standard" rk4 scheme (as provided
by the rk4 solver). It's been forced into the embedded format. In the
future it should be possible to restructure the RKScheme class to allow
such schemes to be better supported, making it easier to implement in a
more efficient way.