Adding a new generic RK/multi-stage solver.
This is currently primarily designed for use with embedded RK schemes
where two orders are used to provide an error estimate. To use select
solver:type=rkgeneric and add solver:scheme=<scheme_type> where
<scheme_type> is one of:
* rkf45 -- Runge-Kutta-Fehlberg embedded order 4/5.
* rkf34 -- Runge-Kutta-Fehlberg embedded order 3/4.
* cashkarp -- Cash-Karp embedded order 4/5.
* rk4simple -- Implementation that matches the existing rk4 solver
  scheme.

It should be relatively simple to add new schemes. These are provided by
a scheme class which essentially defines the order of the scheme, the
number of stages and the Butcher tableau.

There is probably scope to optimise this solver somewhat, particularly
when not adapative (currently we still calculate both order solutions
but just ignore the error estimate).

Don't currently support FSAL optimisation (first same as last -- where
the first stage of a step is identical to the last stage of the previous
step) but should be straight forward to add at a later date.

Whilst the rkscheme base class defines generic routines to construct the
different stages and the resulting solutions, these are virtual and can
be overridden by the specific implementation in order to allow special
treatment/optimisations.