Adding time step factor option.
Can now specify solver:dtfac which multiplies the updated time step when
adaptive. The default for this is 1, but can be overridden in the
derived scheme types.

Also allowing users to force followHighOrder to either true or false to
force the solver to use the high/low order solution respectively.

Making consistency check and Butcher table display hidden unless
solver:diagnose=true