Structural changes in new rk-solver
Moving error calculation into rkscheme class. This allows the generic
solver to be unaware of the alternative order solution (reduces memory
usage) and will allow scheme specific error estimates if desired.

Also restructuring the calculation of the output states to ensure the
second solution is only constructed if we're adaptive. Note that this
results in a small performance hit for the adaptive case. This could be
fixed but will result in messier code.