Minor bug fix to help message.
Previously the handling of the "-h" option meant that any argument
starting "-h" would trigger the help message and then abort the
run. This should now be addressed so the argument must exactly match
either "-h" or "--help".

Note, there's still an outstanding issue that the message is printed
once per processor. To fix this would require us to initialise MPI within the
help message handling (or prior to argument handling).