Unfriending SlepcSolver from Solver
Instead setModel and setRHS are virtual, and intercepted by SlepcSolver.

Options modified, so now gets SLEPc options from the Solver section
passed to the constructor, and then uses a sub-section called "advance"
to specify the time advance solver. This is consistent with
how the PETSc Laplacian solver configures a preconditioner.