Upgrades to SlepcSolver
Two main changes added:

1. Adding support for a custom spectral transform using the shell ST
method. This is currently enabled by adding -st_type shell to the run
line. The advantage of this is that slepc should then deal with the true
BOUT eigenvalue. This means that things like the error calculation and
deciding which eigenvalues are preferred are done directly on Omega
rather than some other function of it.

2. Adding the ability to determine eigenvalues of the ddt operator
rather than the time advance operator. These eigenvalues should be the
same but this new mode is often much faster and does not rely on any
additional solvers. This is now the default operation mode for the
SlepcSolver, to disable add ddtMode = false to the input file.

Minor changes:

1. Making defaults for userWhich and useInitial "smart". In other words
the default for userWhich is true if a target is set and false if not,
the default for useInitial is true unless ddtMode is true (as they don't
seem to work well together).

2. Adding some new comments

3. Made shellMat and slepcToBout members public (required for the shell
wrappers used above).

Some investigation of writing out the eigenmodes during the monitor call
was made but it appears it will not be possible to store the
eigenvectors until the entire EPSSolve call has finished, which prevents
this. We could still write the eigenvalues out as we find them, which
may still be useful, but this is left for a later commit.