Hypnotoad: Adjust last core grid point
If follow_gradient_nonorthog is used to refine
the location of the grid point closest to the
separatrix, this goes wrong close to the X-point
and produces overlapping grids.

This fix adjusts the grid point location to be
close to the intersection of the lines through
core and SOL grid points.