Hypnotoad: Fix for non-orthogonal mesh
Intersection method can go wrong when
the two lines are nearly parallel. This caused
issues with multiple X-points because the point being
refined may not be on the correct separatrix.

Now only uses intersection if the angle is
sharper than 45 degrees. Otherwise uses the standard
LSODE follow_gradient to refine position