Compute intersection with boundaries correctly
Boundaries are between grid points and mesh numbering includes guard
cells, which were not taken into account here. Likewise, the checks
failed to take into account the parallelisation in y (parallelisation in
x,z still not supported).

Lastly, when getting the coordinates of the intersection, the code did
not use the total distance from the boundary point to the field-line
end, nor the direction of the field line.