Subtle bug fix for communicate+TwistShift
The twist-shift parallel boundary condition is applied at the wait stage of
communicate calls. If a fieldGroup contains multiple references to the
same field then during the communicate stage the twist-shift shift can
be applied multiple times which is incorrect (plus communication could
be duplicated).

This commit adds a makeUnique function to the FieldGroup class which is
designed to eliminate any repeat Field references in the
FieldGroup. This is now also automatically called inside communicate in
order to ensure excess communication is avoided and the twist-shift
condition is only applied once.