Add option to increase no. pol pts in pol_slice
This uses interpolation in y (i.e. prior to reconstruction of slice) as
we expect fields to vary slowly in this direction. It would be easy to
extend this to allow interpolation in the other directions.

Can lead to slightly odd looking results near the y boundaries as the
interpolation doesn't include any knowledge of periodicity. May be
possible to improve on this but left for the future.