Large feature commit: Adding adaptivity to imexbdf
Also introduces higher order schemes (up to fourth order).

This commit involves a substantial restructuring of the original
imex-bdf code in order to make it easier to change the order of the
scheme.

Introduced imex-bdf3 and imex-bdf4 schemes.

Adaptivity possible (set adaptive = true). This requires use of
significantly more complicated coefficients (see code for reference) but
general structure of scheme essentially unchanged. Error estimated by
comparing solution with requested order scheme and one order
lower. Rough argument then used to decide how to change the
timestep. This argument works ok for requested order >=3, a little bit
trickier when order = 2.

Some things to note are:
1. Changing the timestep frequently can significantly increase the
number of implicit solves required. The nadapt option aims to help
control this.
2. Recommend order >=3 when adaptive.

Whilst this commit introduces a fully tested and working feature there
are some outstanding elements that are intended to be addressed in the
future:
1. Add documentation of all input parameters (comments in code should be
enough for now).
2. Catch failures of the SNESSolve call to allow a repeat with a smaller
timestep (if desired) rather than aborting as done currently.
3. Develop automatic decision making for how often to check the error
when adaptive.
4. Investigate alternative schemes (e.g. TVB rather than BDF).

Some outstanding issues/concerns also exist:
1. MMS time test (non-adaptive) gives 2nd order for 3rd and 4th order
schemes in some situations (depends on relative size of explicit and
implicit ddt values).
2. Use of two SNES solve object could possibly lead to a drift in
results used for the error checking over time (most likely leading to a
slightly smaller timestep than needed). This is because if the KSP
tolerances are not tight enough the SNES solve can be slightly dependent
on the previous result, which means the high and low order solvers, which
have slightly different solutions from the previous stage, could give
different results.
3. Adaptivity may struggle if the explicit scheme is limiting the
timestep. This isn't tested but could be a slight issue. May want to
introduce a cfl like parameter which allows us to add some safety margin
to our timestep estimate.