Adding writeFlushFreq option to datafile.
This is for use with openclose=true. Rather than opening and closing the
datafile in every call to write, only do it every writeFlushFreq calls
to write. This defaults to 1 in order to give the standard
behaviour. Note this will not be useful for many runs but can help when
you would like high time resolution output (e.g. when I/O is taking a
significant amount of each output step).

Note you are likely to get jumps in the solver time every writeFlushFreq
outputs (it looks like this write time is actually ending up being
attributed to solver rather than I/O).