Fix for use with intel compilers without openmp.
For some reason the references used to store *min and *max in the data
iterator were being set to the wrong memory locations when using intel
17 compilers. Using separate int vars (as done when running with openmp)
seems to fix this problem. Whilst this removes the nicely explicit
behaviour that without openmp *min should be == *start etc. it at least
means we get the correct behaviour with intel and there's a unified
declaration of these vars.