Bugfix: FieldGroup handles Vectors
Vector2D and Vector3D objects could be added to FieldGroup, but
would then cause a segfault when communicated. This is because the
communication code assumes the objects are either Field3D or Field2D.

This fix catches Vector2D and Vector3D additions to FieldGroup,
converting them into Field2D or Field3D components.