Output settings to file
The options used by BOUT++, including any overwritten by
command-line arguments, are saved to a file BOUT.settings

This is done twice:
* Once during initialisation, so that the options are saved
  if the simulation stops/crashes
* Again at the end, so the comments include whether the option
  was used during the simulation.

The format used is the same as the BOUT.inp file, so in
principle it should be possible to rename BOUT.settings to BOUT.inp
to run the simulation again.

Note that this does not capture command-line arguments to libraries
such as PETSc.

This fixes issue #451, and is a partial fix for issue #35 and issue #233