Moving restart logic from Solver to PhysicsModel
The solver is no longer responsible for loading and saving
state to restart files. Instead this is done in PhysicsModel.

A function postInit() is run after init(). This loads the restart
file, overwriting any values set in init. If a user wants different
behavior, they can implement a differen postInit() function.

Restart file writing is currently done in runOutputMonitor,
so cannot be modified by the user. In principle though this could be
moved to a separate function so that the user could override
this behavior.

This fixes issue #439