Adding nyInterp and withRepeat options to pol_slice
The withRepeat input can be used to indicate if the data file includes
the duplicate z grid point (e.g. for compatibility with bout++ < v4).

The nyInterp input can be used to request interpolation in the
y (parallel) direction prior to constructing the poloidal cross
section. The idea here is that we tend to expect things to be slowly
varying in the parallel direction so interpolation should be reasonable,
and we often want more points in the poloidal direction than actually
used in the simulation for nice figures.