FieldGroup catch incorrect types at compile time
If FieldGroup was constructed using a type which could not be added,
such as an integer or const Field3D, then an infinite loop would result.
This is because the call to add() inside the constructor would try to
implicitly convert the type to FieldGroup, using the variadic template
constructor.

This fixes issue #540 by making the template constructor explicit.

The infinite loop no longer occurs since the call to add() inside the constructor
fails to compile. Instead an error message like this is printed:

    ../../include/bout/fieldgroup.hxx: In instantiation of ‘FieldGroup::FieldGroup(Ts& ...) [with Ts = {int}]’:
    test_fail.cxx:14:17:   required from here
    ../../include/bout/fieldgroup.hxx:57:36: error: no matching function for call to ‘FieldGroup::add(int&)’
        explicit FieldGroup(Ts&... ts) { add(ts...); }

A test case has also been added (examples/test-fieldgroup) which tests for this issue
and a couple of other sanity checks.