Changing mesh:symmetricGlobalY to true by default
This is the "correct" setting, which corresponds to setting y
to 0 and 2pi on the boundaries in the input expressions.

Some test cases depend on the original setting,
so now set symmetricGlobalY to false in the BOUT.inp settings.

This changes user inputs, but if anyone has been doing testing
(particularly MMS), they will already have been setting symmetricGlobalY
to true (or getting frustrated trying to work out why their test
doesn't converge properly).

This fixes issue #484