Configure finds gmake
If the gmake command exists then use that rather than make.
This is because the makefiles work with GNU make but not BSD make.

MAKE can be set on the configure command line to override

The make command is now printed at the end of the configure
script, telling the user what to run to compile the code.