Allows the TRACE macro to take additional arguments used for formatting
In other words `TRACE("Some string %d",2)` would lead to a `msg_stack`
entry "Some string 2 on line __LINE__ of '__FILE__'". This means `TRACE`
has the full capability of `msg_stack.push`.

To handle the cases with and without additional arguments need to either
use a gcc pre-processor extension *or* do some (possibly ugly) magic. To
remain flexible I've gone for the magic route but easy to switch if required.