Simplifying TRACE macro for variable argument count.
Requires a rearrangement of the arguments in the MsgStackItem
constructor for the case with a variable argument list.

Two original constructors kept currently but are not used anywhere in
the BOUT++ code base so should perhaps be removed. If kept for long term
then may want to consider reordering their arguments as well for consistency.