Replacing modulus operation with ternary
inside triple nested loop used for Arakawa bracket
calculations. These operations are for finding `jz+1` and `jz-1` and
wrapping these around. Profiling has found that this can reduce the
time spent in the bracket operation by a factor of two on modern hardware.

The compiler could not make this optimisation as it doesn't know that we
can only ever exceed the limits by up to 1 (i.e. so we know the result
of the modulus is only ever going to be the input or  `0`, effectively)