Adding new_names and metric2d options to write_maps
new_names causes write_maps to output "g_xx", "g_yy" etc.
rather than "g_11", "g_22" etc.

metric2d reduces the 3D metrics to 2D, by taking only the Z=0
part and dz(0,0,0). This is to allow it to write grids
compatible with the current BOUT-dev. No checks are currently
performed that this is ok, and it should only be done for
rectangular grids.