Adding header file that provides some wrappers around scorep
routines for user instrumentation. To be active requires
`BOUT_HAS_SCOREP` to be defined (need to add this to configure).

Need to add a test to configure to see if the compiler being used
supports `__PRETTY_FUNCTION__` (should be supported by gcc, clang and
intel at least). If so then we should define `__HAS_PRETTY_FUNCTION`.

Finally the "level" of the instrumentation to be used is controlled by
`SCOREPLVL`, set negative to disable all instrumentation.