ParderivFactory now checks for type
Previously the default Create() (should be create...)
would create a serial type if NYPE == 1, and a cyclic type
otherwise.  Unfortunately the serial type doesn't work
on open flux surfaces, and is mainly there for testing.

This resulted in unnecessary errors when running some SD1D
cases (e.g. case-05) on one processor.

The cyclic type works on one processor, so this check
of NYPE is not needed. Instead the factory now gets
the "parderiv:type" option, which defaults to cyclic.