Default options are inserted into Options tree
The first time a value is requested which has not been set,
the default value is put into the options tree with the source
"default". This means that these options are written to the
BOUT.settings file.

Subsequent calls requesting the same value check if the default
values are consistent. If default settings are different then
an exception is thrown.