Turning off logging of options to BOUT.log by default
There is now a "-v" or "--verbose" flag which turns back on logging
of options to BOUT.log files.

The log flag can now be set for each Options section
separately, allowing logging for parts of the code if desired.
Changes affect sections and all subsections, so
changes to the root Option object affect all options.