Solver: Making args to load_vars, load_derivs const
This is needed to handle changes in imexbdf2 solver, in turn
driven by changes in PETSc. These functions don't modify
their arguments, so can be const.

An issue is that the loop_vars function (which should be replaced)
does not distinguish between const/non-const. For now I have just
used a const_cast to remove const when passing to loop_vars.