IMEX-BDF2 changing VecGetArray to VecGetArrayRead
A recent PETSc version (on or before 3.7.6) added VecLock
functions, which enforce read-only vectors. IMEX-BDF2 was
using VecGetArray on input vectors which should not be modified,
resulting in a runtime error.

This changes VecGetArray to VecGetArrayRead for input arrays,
fixing the PETSc errors.

Since VecGetRead was added in PETSc 3.4.0 there is a compile
switch #if PETSC_VERSION_GE(3,4,0) around this.