Replacing static arrays with Array
Several places in the code use arrays for intermediate results,
which were previously resized as needed. This was not really
a memory leak, because everything would be cleaned up at the end
of the simulation. It was however not thread safe, and made the code
harder to read than needed.

Removing unused invert/parderiv/impls/simple