Adding `invalidateGuards` routines for 3D and 2D
This routine sets the guard cells in x and y to be `nan("")`.

The general idea is that we can use these in routines when we know that
the guard cells no longer hold the correct values (e.g. after
differential operations etc.). If employed sufficiently widely this
might help people track down missing communicates and similar. It also
provides a form of documentation -- we're clearly indicating that this
routine invalidates the guard cells.