Some optimisations for operators_di
Making values and indices const in bracket_arakawa

Moving the Mesh::coordinates function into the header.
The Coordinates object can't be created there unfortunately,
because Coordinates depends on Mesh. Instead this is put in
a private Mesh method.

Doesn't seem to make a noticeable difference on laptop
with -O3 (core i3-3227U)