First stab at removing free_boundary treatment in mesh
This removes the special handling of boundary regions in `apply?Diff`
routines for free boundary code and greatly simplifies this code.

It should be noted that there is now a `free` boundary option which
should still allow this behaviour to be used but has several
benefits (such as being per field rather than set at the mesh level).

THIS SHOULD BE CONSIDERED EXPERIMENTAL, NOT MUCH THOUGHT PUT IN.