install paths in make.config and bout-config
New variables BOUT_INCLUDE_PATH, BOUT_LIB_PATH and BOUT_CONFIG_FILE
defined in make.config and bout-config files. These point to the
include paths for cflags, ldflags, and the make.config file.

During configure these variables are set to the working directory,
so the current process of compiling and running in the same directory
still works.

When "make install" is run, these paths are replaced in the installed
scripts, to point to the new locations.

The bout-config script "--prefix" option has been replaced by
"--config-file". This returns the full name of make.config.
A user makefile can be:

    SOURCEC = myfile.cxx

    include $(shell bout-config --config-file)

Alternatively, a more flexible but longer way is to use --cxx (compiler),
--cflags, --ld (linker), --libs

This method captures the flags pulled in by external codes like PETSc,
using the output.make targets in configure. If the make.config method
is used then the path to PETSc (for example) should be consistent with
the path used in configure.

The paths to the IDL and Python libraries can also be accessed using
bout-config --idlpath and --pythonpath to set up environment variables.

Also added a "make uninstall" target, which takes the same DESTDIR
argument as install, and deletes the installed files.