Removing Field3D move constructor
Relates to issue #698, but is only a temporary fix.

The "default" move constructor seems to do something
involving the yup/down fields which can in some cases lead
to segfaults.

The longer-term solution is probably to implement a move
constructor which takes into account pointers to
yup/down fields etc.