Add an option "ddz:fft_filter" to control filtering in DDZ
This partly fixes issue #647. A variable "fft_derivs_filter" in mesh
is initialised using an input option. This is then used in the FFT
derivative function to control how many modes are filtered out.

Note: This changes the default, from filtering 60% of the modes to
filtering nothing. I think the correct default should be to not filter
unless the user asks for filtering. I think this is now the case
elsewhere where Fourier filtering can appear.