test-yupydown uninitialised values. Mostly harmless.
filename buffer in DataFile was not initialised, leading to valgrind
warning. In practice this buffer was always written to before being
used, but valgrind didn't like the move assignment.

In test-yupydown some of the uninitialised data warnings seem to be
false positives, since they go away when the return field is first
set to zero then overwritten.

Still have a persistent memory write error/warning in test-yupydown
during communicate:

==1549== Invalid write of size 8
==1549==    at 0x4C33C33: memmove (in /usr/lib/valgrind/vgpreload_memcheck-amd64-linux.so)
==1549==    by 0x1592708F: ??? (in /usr/lib/openmpi/openmpi/mca_btl_self.so)
...
==1549==    by 0x21AE2D: BoutMesh::send(FieldGroup&) (boutmesh.cxx:881)
...
==1549==  Address 0x1562ca40 is 0 bytes inside a block of size 12,288 alloc'd
==1549==    at 0x4C2DC9F: operator new[](unsigned long) (in /usr/lib/valgrind/vgpreload_memcheck-amd64-linux.so)
==1549==    by 0x21F4F9: BoutMesh::get_handle(int, int) (boutmesh.cxx:1828)

The block it's inside is dmesg_recvbuff, but the buffer being sent is
umsg_sendbuff. Making these buffers 10 times larger didn't remove the
warning. This may be something strange in OpenMPI, or maybe not...